+++ client/src/node/extension.ts
@@ -6,8 +6,6 @@ import {
   NotebookCellData,
   NotebookCellKind,
   NotebookData,
-  StatusBarAlignment,
-  StatusBarItem,
   Uri,
   authentication,
   commands,
@@ -41,8 +39,13 @@ import { installCAs } from "../components/CAHelper";
 import ContentNavigator from "../components/ContentNavigator";
 import { setContext } from "../components/ExtensionContext";
 import LibraryNavigator from "../components/LibraryNavigator";
-import { LogTokensProvider, legend } from "../components/LogViewer";
 import ResultPanelSubscriptionProvider from "../components/ResultPanel";
+import {
+  getStatusBarItem,
+  resetStatusBarItem,
+  updateStatusBarItem,
+} from "../components/StatusBarItem";
+import { LogTokensProvider, legend } from "../components/logViewer";
 import { NotebookController } from "../components/notebook/Controller";
 import { NotebookSerializer } from "../components/notebook/Serializer";
 import { ConnectionType } from "../components/profile";
@@ -50,14 +53,12 @@ import { SasTaskProvider } from "../components/tasks/SasTaskProvider";
 import { SAS_TASK_TYPE } from "../components/tasks/SasTasks";
 
 let client: LanguageClient;
-// Create Profile status bar item
-const activeProfileStatusBarIcon = window.createStatusBarItem(
-  StatusBarAlignment.Left,
-  0,
-);
+
+export let extensionContext: ExtensionContext | undefined;
 
 export function activate(context: ExtensionContext): void {
   // The server is implemented in node
+  extensionContext = context;
   const serverModule = context.asAbsolutePath(
     path.join("server", "dist", "node", "server.js"),
   );
@@ -89,8 +90,6 @@ export function activate(context: ExtensionContext): void {
     serverOptions,
     clientOptions,
   );
-
-  activeProfileStatusBarIcon.command = "SAS.switchProfile";
 
   // Start the client. This will also launch the server
   client.start();
@@ -129,14 +128,14 @@ export function activate(context: ExtensionContext): void {
     authentication.registerAuthenticationProvider(
       SASAuthProvider.id,
       "SAS",
-      new SASAuthProvider(context.secrets),
+      new SASAuthProvider(),
     ),
     languages.registerDocumentSemanticTokensProvider(
       { language: "sas-log" },
       LogTokensProvider,
       legend,
     ),
-    activeProfileStatusBarIcon,
+    getStatusBarItem(),
     ...libraryNavigator.getSubscriptions(),
     ...contentNavigator.getSubscriptions(),
     ...resultPanelSubscriptionProvider.getSubscriptions(),
@@ -165,19 +164,20 @@ export function activate(context: ExtensionContext): void {
   );
 
   // Reset first to set "No Active Profiles"
-  resetStatusBarItem(activeProfileStatusBarIcon);
+  resetStatusBarItem();
   // Update status bar if profile is found
-  updateStatusBarProfile(activeProfileStatusBarIcon);
+  updateStatusBarItem();
 
   profileConfig.migrateLegacyProfiles();
   triggerProfileUpdate();
 }
 
 function triggerProfileUpdate(): void {
+  commands.executeCommand("SAS.close", true);
   const profileList = profileConfig.getAllProfiles();
   const activeProfileName = profileConfig.getActiveProfile();
   if (profileList[activeProfileName]) {
-    updateStatusBarProfile(activeProfileStatusBarIcon);
+    updateStatusBarItem();
 
     const connectionType =
       profileList[activeProfileName].connectionType || ConnectionType.Rest;
@@ -200,38 +200,6 @@ function triggerProfileUpdate(): void {
       ConnectionType.Rest,
     );
   }
-}
-
-async function updateStatusBarProfile(profileStatusBarIcon: StatusBarItem) {
-  const activeProfileName = profileConfig.getActiveProfile();
-  const activeProfile = profileConfig.getProfileByName(activeProfileName);
-  if (!activeProfile) {
-    resetStatusBarItem(profileStatusBarIcon);
-  } else {
-    const statusBarTooltip = profileConfig.remoteTarget(activeProfileName);
-
-    updateStatusBarItem(
-      profileStatusBarIcon,
-      `${activeProfileName}`,
-      `${activeProfileName}\n${statusBarTooltip}`,
-    );
-  }
-}
-
-function updateStatusBarItem(
-  statusBarItem: StatusBarItem,
-  text: string,
-  tooltip: string,
-): void {
-  statusBarItem.text = `$(account) ${text}`;
-  statusBarItem.tooltip = tooltip;
-  statusBarItem.show();
-}
-
-function resetStatusBarItem(statusBarItem: StatusBarItem): void {
-  statusBarItem.text = `$(debug-disconnect) ${l10n.t("No Profile")}`;
-  statusBarItem.tooltip = l10n.t("No SAS Connection Profile");
-  statusBarItem.show();
 }
 
 export function deactivate(): Thenable<void> | undefined {