+++ server/src/server.ts
@@ -1,8 +1,7 @@
-// Copyright © 2022, SAS Institute Inc., Cary, NC, USA.  All Rights Reserved.
+// Copyright © 2022-2023, SAS Institute Inc., Cary, NC, USA.  All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 import {
   Connection,
-  DidChangeConfigurationNotification,
   InitializeResult,
   SemanticTokensRequest,
   TextDocumentSyncKind,
@@ -24,19 +23,6 @@ let supportSASGetLibList = false;
 // Create a simple text document manager.
 const documents: TextDocuments<TextDocument> = new TextDocuments(TextDocument);
 
-// Add settings wanted to listen here, onInitialized(), and onDidChangeConfiguration()
-interface ListenedSettings {
-  editor?: {
-    tabSize?: number;
-    insertSpaces?: boolean;
-  };
-}
-const cachedConfigurations: ListenedSettings = {};
-
-// Default setting values
-const DEFAULT_TAB_SIZE = 4;
-const DEFAULT_INSERT_SPACES = true;
-
 export const init = (conn: Connection): void => {
   connection = conn;
   connection.onInitialize((params) => {
@@ -54,6 +40,7 @@ export const init = (conn: Connection): void => {
           full: true,
         },
         documentSymbolProvider: true,
+        documentFormattingProvider: true,
         foldingRangeProvider: true,
         hoverProvider: true,
         completionProvider: {
@@ -64,29 +51,12 @@ export const init = (conn: Connection): void => {
           firstTriggerCharacter: "\n",
           moreTriggerCharacter: [";"],
         },
+        signatureHelpProvider: {
+          triggerCharacters: ["(", ","],
+        },
       },
     };
     return result;
-  });
-
-  connection.onInitialized(() => {
-    // Initialize listened settings
-    connection.client.register(DidChangeConfigurationNotification.type, {
-      section: ["editor"],
-    });
-    connection.workspace
-      .getConfiguration([
-        {
-          section: "editor",
-        },
-      ])
-      .then((data) => {
-        cachedConfigurations.editor = data[0];
-      });
-  });
-
-  connection.onDidChangeConfiguration((params) => {
-    cachedConfigurations.editor = params.settings.editor;
   });
 
   connection.onRequest(SemanticTokensRequest.type, (params) => {
@@ -114,37 +84,56 @@ export const init = (conn: Connection): void => {
 
   connection.onDocumentSymbol((params) => {
     const languageService = getLanguageService(params.textDocument.uri);
-    return languageService
-      .getFoldingBlocks()
-      .filter((symbol) => symbol.name !== "custom");
+    return languageService.getDocumentSymbols();
   });
 
   connection.onFoldingRanges((params) => {
     const languageService = getLanguageService(params.textDocument.uri);
-    return languageService.getFoldingBlocks().map((block) => ({
-      startLine: block.range.start.line,
-      endLine: block.range.end.line,
-    }));
+    return languageService.getFoldingRanges();
   });
 
   connection.onRequest("sas/getFoldingBlock", (params) => {
     const languageService = getLanguageService(params.textDocument.uri);
-    return languageService.getFoldingBlock(params.line, params.col);
+    const block = languageService.getFoldingBlock(
+      params.line,
+      params.col,
+      params.strict ?? true,
+      params.ignoreCustomBlock,
+      params.ignoreGlobalBlock,
+    );
+    if (!block) {
+      return undefined;
+    } else {
+      return { ...block, outerBlock: undefined, innerBlocks: undefined };
+    }
   });
 
   connection.onDocumentOnTypeFormatting((params) => {
-    const tabSize: number =
-      cachedConfigurations.editor?.tabSize ?? DEFAULT_TAB_SIZE;
-    const useSpace: boolean =
-      cachedConfigurations.editor?.insertSpaces ?? DEFAULT_INSERT_SPACES;
     const languageService = getLanguageService(params.textDocument.uri);
     return languageService.formatOnTypeProvider.getIndentEdit(
       params.position.line,
       params.position.character,
       params.ch,
-      tabSize,
-      useSpace,
+      params.options.tabSize,
+      params.options.insertSpaces,
     );
+  });
+
+  connection.onSignatureHelp((params) => {
+    const languageService = getLanguageService(params.textDocument.uri);
+    completionProvider = languageService.completionProvider;
+    return completionProvider.getSignatureHelp(
+      params.position,
+      params.context?.activeSignatureHelp?.activeSignature,
+    );
+  });
+
+  connection.onDocumentFormatting((params) => {
+    const languageService = getLanguageService(params.textDocument.uri);
+    return languageService.formatter.format({
+      tabWidth: params.options.tabSize,
+      useTabs: params.options.insertSpaces === false,
+    });
   });
 
   documents.onDidChangeContent((event) => {